chore(ci): test http proxy used in request to check for new app versions
